Working on interpolation (added edge wrapping) Fixed attr bug caused by speed being an float
git-svn-id: https://svn.torandi.com/rocket@49 71c61736-3055-43ec-afd9-7e59d617a297

+++ c_client/ship.h
@@ -9,7 +9,7 @@ struct ship_t {
 	int x;
 	int y;
 	int a; //angle
-	int s; //speed
+	float s; //speed
 	bool attr[NUM_GFX_ATTR];
 	
 	int _x;
